requirements.txt
asyncpg
gymnasium
pygame
torch
# torch from: https://pytorch.org/get-started/locally/
torchvision
# torchvision from: https://pytorch.org/get-started/locally/
tqdm
onnxruntime
onnx
pandas
requests
Pillow
numpy
